Add a Makefile option to disable libunwind
There's been a request for a version of julia that does not depend on
libunwind, for easier bring up on new architectures. This adds a makefile
flag that can be put in Make.user (`DISABLE_LIBUNWIND`) and if set, will
disable any dependency on libunwind. Backtraces and profiling won't work
of course.